third_party/blink/web_tests/media/controls/large-overflow-menu-when-pip-enabled.html

<!DOCTYPE html>
<title>Large overflow menu when pip enabled</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/run-after-layout-and-paint.js"></script>
<script src="../media-controls.js"></script>
<body>
<video controls></video>
<script>
async_test(t => {
  let video = document.querySelector("video");
  video.src = "../content/test.ogv";
  enableTestMode(video);

  const overflowMenuList = overflowMenu(video);
  const textTrackList = textTrackMenu(video);

  assert_true(document.pictureInPictureEnabled, 'pip should be enabled');

  video.onloadedmetadata = t.step_func(() => {
    expectOverflowMenuAndTrackListContainsPip();

    video.setAttribute('disablepictureinpicture', '');
    runAfterLayoutAndPaint(t.step_func(() => {
      expectOverflowMenuAndTrackListNotContainsPip();

      video.removeAttribute('disablepictureinpicture');
      runAfterLayoutAndPaint(t.step_func_done(() => {
        expectOverflowMenuAndTrackListContainsPip();
      }));
    }));
  });

  function expectOverflowMenuAndTrackListContainsPip() {
    assert_true(overflowMenuList.classList.contains('pip-presented'), 'overflow menu should have pip-presented');
    assert_true(textTrackList.classList.contains('pip-presented'), 'track list should have pip-presented');
  }

  function expectOverflowMenuAndTrackListNotContainsPip() {
    assert_false(overflowMenuList.classList.contains('pip-presented'), 'overflow menu should not have pip-presented');
    assert_false(textTrackList.classList.contains('pip-presented'), 'track list should not have pip-presented');
  }
});
</script>
</body>
